fix(cookieService): properly close the file
The old version of this routine would leave a bunch of dangling file
handles because it used `defer` to close the cookiejar. However, the
`cookieService` goroutine never exited, it just went to sleep and looped
over again. This force closes the cookiejar after reading it, preventing
an explosion of file handles.

Closes #1

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>