feat(request): Allow automatic request of cookies
This method introduces a very basic and specific method to query cookies
from a remote server and to save those cookies. This eliminates the need
for curl running beside CookieProxy to fetch some types of cookies. It's
really a hyperspecific configuration just for the Tesla Powerwall API,
but it's going to fill my need. In the future I'll add in addtional
options like allowing for username/password to other hosts and maybe
even allowing multiple requests to cache a wider variety of cookies.

This is really specific to my use case around the Tesla Powerwall API,
but it works well enough for me right now.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>