feat(debug): add a debug flag for more output
When using cookieproxy with the `-request` option, you may not want to
see the complete request as it likely has sensitive data. This makes it
so that information is hidden unless you use `-debug`.

This also fixes a JSON parsing bug that I uncovered while testing.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>